platform/atrace-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atrace-orchestrator"
version = "0.1.0"
description = "Task orchestration layer for TraceMind — pipeline runner, rule engine, reporter"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = ["pydantic>=2.0.0", "pyyaml>=6.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["atrace_orchestrator*"]

[tool.setuptools.package-data]
"atrace_orchestrator" = ["playbooks/*.yaml"]
